refactor(rust): extract system.rs with utility commands
